Use esy and patch Ligo to make it build on Windows.
Problem with paths is it is assumed Windows prefers \ but because of
all the Unix-ism's in the OCaml ecosystem, this breaks things.

Why module name shortening.
Dune creates long paths based on a libraries' name (the name field in
dune files). Despite having long paths enabled in the Windows
registry, Dune isn't able to rename/mv files that have long paths. The
workaround I use to to simply shorten the library name. This
workaround is dependent on the machine's $HOMEPATH - the CI's is
different from local one. So, while during the iteration where I build
things locally, I could fix the library names that are problematic,
but on the CI, in addition, we'll need more.