[LIGO-833] Display partially applied arguments
Problem: it would be nice if for partially applied functions we
displayed what has been applied and what is the original function.

Solution: generalize `LambdaMeta` to carry such information.
It became more complex, but I also believe that for the new data model
association with what is going on in LIGO is clearer.

Update variables rendering to allow for partial application trees like
+ add8: <lambda>
  + func: add5
    + func: add
    + arg0: 5
  + arg0: 3