[LIGO-833] Add a test to `test/Snapshots.hs`
Problem: we tested `buildLambdaInfo`, but didn't yet test how
`LambdaMeta` is constructed.

Solution: add a test next to tests on snapshots logic, as our logic is
quite related to stack frames.

Use `advanced-curry.mligo` contract as a quite complex case, and update
it to include a zero-arg partial application (basically a renaming).
Such complex case should be quote enough as the logic for forming
`LambdaMeta` is quite simple.

Also that contract was updated to `result = applyThrice d + 1` (added `+
1`), in the current reality this is necessary if we want to create a
snapshot at that line and investigate the functions' internals.

Add two tests:
* That appropriate `LambdaMeta` is constructed.
* That this meta results in proper variables (sort of an integrational
  test).