Make tezos-ligo src as Dune vendored deps again
Even if pin-depends field in the opam manifest reads the opam manifest
of tezos-* deps from ./vendors/tezos-ligo, it doesn't really install
anything meaningful. It only installs deps of the tezos-*
packages (not the packages themselves). opam thinks opam/
folder (inside tezos-ligo) is the source tree and simply copies it to
install location. The packages are installed by Dune.

This behaviour is different in esy where it correctly guesses the
source tree and installs the artifacts. This is how we're able to work
around the long paths issue on Windows - we don't have to install
tezos-*  dependencies as dune vendored deps (which are built in a long
path location). esy uses it's short path location ~/.esy/3/b to build
the artifacts.