[LIGO-833] Add safe `Buildable` instance for `LigoType`
Problem: in tests in for logging it would be convenient to have
`Buildable` instances for `LambdaMeta` and a bunch of related types, but
they contain `LigoType` that lacks `Buildable` instance.

The reason for this absence of instance - pretty printing depends on
the used dialect. If we use some default dialect and that leaks to the
user's output, that would be hard to notice and actually quite a severe
issue.

Solution: add `Buildable LigoType` instance with `ForInternalUse` dummy
constraint.

This constraint is supplied by hand when the developer is sure that the
context is appropriate; e.g. `logMessage` provides this constraint into
the message.

I'm not sure this solution is the best possible one, but it should not
be hard to change if we decide something else fits better.