[#1546] Bring back LIGO-474 regression test
Problem: Our previous `treeDoesNotContainNameTest` used a ReasonLIGO
contract that was deleted. So we need a substitute.

Solution: Get a contract from the error-recovery tests that fail in the
same manner and use it instead. Restore deleted tests.

n.b.: This test has a pretty ugly message, we should handle this case
better.