[#1532] Add delayed variables computation
Problem: we are about to implement displaying values in LIGO format,
which requires calling `ligo` executable.

And this may take long, there is still no way to run commands against
`ligo` quickly.

Solution: allow delayed value computation.

When VSCode first asks for the variables, we return
nothing (`computing...` stub) and start evaluating all the requested
variables in batch. Once the result comes in, we invoke
`InvalidatedEvent` to ask VSCode re-request the variables.

Variables are stored in a dedicated abstraction that adds a memoization
to a `IO` function, so once a given Michelson value of a given LIGO type
in given LIGO dialect is computed, we remember the result forever and
on later computation attempts for the same value we return the result
instantly.

Keeping the result forever seems fine, since we remember all the
information about past snapshots anyway, and it is no smaller.

I embed the new cached computation logic into the DAP Handler.
Snapshots generation logic, as well as the stages prior to it are not
affected.
Motivation: it's better keep snapshots generation logic pure and close
to the Michelson interpreter, and all derivative data is better be
added at later stages rather than directly in the `Snapshots.hs` module
for the sake of separation of concerns.

Alternative approaches:

I first tried to keep some `TVar` with a value right in `StackItem`s,
but it spoiled too much things - `Eq` and `Show` instances for
everything in `Snapshots.hs` flew away and that would be a suffering for
the test code.

Moreover, the currently applied approach is better in that most of the
variables live for at least a few snapshots, and we won't compute it
unnecessarily after each single step.