[#1532] Deal with the computation threads
Problem: currenty variables computation threads are spawn with `async`
and the resulting reference to the thread is thrown away, and this is
never a good thing.

Solution: what we want here is to:
1. Remember the spawned threads.
2. Kill them on debug session termination.
3. Make sure the number of the spawned threads is capped.

So I add an abstraction that acts like a thread pool but once reaching
the upper bound on the threads number it aborts the older threads before
spawning new ones. This way we ensure that we do not keep dozens of
`ligo` processes simultaneously.