[LIGO-833] Make curried funs displayed with LIGO values
Problem: the previous commits stopped working at some point, since when
I try to build lambda's children variables, the original lambda is
already replaced with LIGO value. Moreover, this LIGO value is not a
lambda but a `<fun>` string.

Solution: add the original Michelson value to `LigoValue` constructor of
`LigoOrMichValue`, and in `buildSubVars` use that to get the necessary
lambda meta.

This solution has several drawbacks:

1. Ideally we have to zip LIGO value with the Michelson value, the
current solution doesn't do this which means the feature is working only
for lambdas that are at top-level of the `VARIABLES` pane, and not
within e.g. a product type.

However doing the zip is very non-trivial - product types have different
layouts and this must be accounted. I leave doing so for future work.

Ideally, perhaps, we have to somehow pass the lambda meta through the
decompilation process, but this has its own caveats.

2. The fact that the feature fires after replacing Michelson lambda with
`"<fun">` and after replacing that with LIGO variable - seems
conceptually wrong. Not like it should cause bugs, but maybe we have to
better organize our variables processing pipeline. Not sure how though.