[#1888] Fix disappearing variables in the debugger
Problem: if the variable is unused in the further stepping then it
disappears from the variables pane. Sometimes it becomes annoying because
it requires to do some step backs to see this variable again.

Solution: suppress strengthening pass in the compiler when optimizations
are turned off.

:NOTE: this suppression breaks shadowing. Now we can see all instances of
the same named variables. It's questionable how to fix this (or should we?).